# docs/migration/contracts/api.yaml
...
openapi: 3.1.0
info:
  title: Silicon Extend API
  version: 2.0.0-draft
  description: |
    Silicon Extend lets a Silicon use a Carbon's personal devices. This API is used by the `extend` CLI
    (through the `silicon-extend-client` crate), the configuration website, and the Extend apps on devices.

    Conventions
    - Every JSON body is an envelope: `{"type": "<kind>", "data": {...}}`.
    - Errors: `{"type": "error", "data": {"code", "message", "hint", "docs_url", "request_id", "details"}}`.
      `code` is stable and machine-readable; `message` says what happened and why; `hint` says what to do next.
    - Timestamps are RFC 3339 UTC with milliseconds. Durations carry their unit in the field name.
    - Every response carries `X-Request-ID` and `Silicon-Extend-API-Version`.
    - Lists are paged with `limit` and `cursor`; the response gives `next_cursor`, null on the last page. A page that
      has a `next_cursor` is always full; only the last page can be short.
    - Retry a POST after an uncertain result with the same `Idempotency-Key` and the identical body. The same key with
      a different body is `409 conflict`.
    - API majors (TECHNICAL.md section 10). Every `/api/v{n}/` request is checked before anything else: a major this
      service doesn't serve is `400 api_version_unsupported`; a `Silicon-Extend-API-Version` pin naming another major
      than the path is `400 api_version_mismatch`; a sunset major is `410 api_version_sunset`, whose hint says how to
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /webhooks/accounts:
    post:
      tags: [Accounts webhooks]
      operationId: accountsWebhook
      summary: Silicon Accounts' signed events about the accounts that use Extend
      description: |
        `X-Accounts-Signature: v1=<hex HMAC-SHA256>` over `"{X-Accounts-Timestamp}.{raw body}"` with
        `EXTEND_ACCOUNTS_WEBHOOK_SECRET` (then `EXTEND_ACCOUNTS_WEBHOOK_PREVIOUS_SECRET` during a rotation), 5 minutes of
        tolerance, verified over the exact raw body before anything is read: otherwise `401`. A signed body that isn't
        an event is `400`. Deliveries are deduplicated on `event_id`; effects are idempotent, and a failed one answers
        `500` so Silicon Accounts retries it. Types Extend doesn't act on are acknowledged (`204`).

        - `account.id_changed`, `account.updated` (in version order): the cached id, name and photo; everything stays
          keyed on the uuid.
        - `silicon.custodian_changed`: the new custodian sees the Silicon; the grants the previous custodian gave it
          end (sessions there end `access_removed`); other Carbons' grants stay.
        - `membership.signed_out` (except `app_revoked`) and `membership.access_removed`: tokens issued before the
          event are refused; a Silicon's sessions end (`silicon_logged_out`), a Carbon's Silicons' sessions on their
          pairs end (`access_removed`); proofs Extend held for the account are revoked.
        - `membership.signed_out` with `app_revoked` (Extend revoked one sign-in: a logout): what
          `POST /api/v2/auth/logout` ends, for sessions started before the logout; other sign-ins stay valid.
        - `account.deleted`: as above, and a Carbon's pairs end (the devices are unpaired), a Silicon's grants end
          and its wake requests are withdrawn; others' history shows "deleted account".
      security: []
      parameters:
        - {name: X-Accounts-Timestamp, in: header, required: true, schema: {type: string, example: '1790000000'}}
        - {name: X-Accounts-Signature, in: header, required: true, schema: {type: string, example: 'v1=5d41…'}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, type]
              properties:
                event_id: {type: string}
                type: {type: string, example: membership.signed_out}
                occurred_at: {type: string, format: date-time}
                app_id: {type: string, example: extend}
                data: {type: object}
      responses:
        '204': {description: 'Applied, already applied, or a type Extend doesn''t act on'}
        '400': {description: 'invalid_input: not a Silicon Accounts event', $ref: '#/components/responses/Error'}
        '401': {description: 'unauthorized: bad, missing or stale signature', $ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Silicon Accounts access token for the app `extend`: an EdDSA-signed JWT whose `aud` is `extend` and whose
        `iss` is the Silicon Accounts this server trusts (`ACCOUNTS_URL`), verified with its published keys. Carbons get
        one with the Silicon Accounts device flow (the CLI) or the authorization-code flow (the website's server),
        Silicons by exchanging a short-lived token from `silicon-accounts login --app extend -q`. Tokens issued before
        a sign-out, removed access or deleted account are refused.
    enrollmentSecret:
      type: http
      scheme: bearer
      bearerFormat: 'Enrollment secret (ees_). Sent as "Authorization: Extend-Enrollment <secret>".'
    deviceCredential:
      type: http
      scheme: bearer
      bearerFormat: 'Device credential (edc_). Sent as "Authorization: Extend-Device <credential>".'
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [type, data]
            properties:
              type: {const: error}
              data: {$ref: '#/components/schemas/Error'}
          example:
            type: error
            data:
              code: device_in_use
              message: 'Device 7c1e09ab (Saket''s Pixel) is being used by si:chef in session a3f since 2026-09-26T09:58:00Z. Only one Silicon can use a device at a time.'
              hint: 'Ask for it with: extend request send 7c1e09ab --reason "<why, up to 300 characters>"'
              docs_url: https://extend.teamofsilicons.com/docs/cli#error-device_in_use
              request_id: 01926f41-0a1b-7c2d-8e3f-4a5b6c7d8e9f
              details: {in_use: {silicon_id: 'si:chef', session_id: a3f, since: '2026-09-26T09:58:00.000Z'}}
          # When the Silicon using the device is on another side (another Carbon's pair, or outside the caller's custodian
          # circle), Extend names no one: "Device 7c1e09ab (Living room TV) is in use. Only one Silicon can use a device
          # at a time.", hint 'Ask for it: extend request send 7c1e09ab --reason "..."', details {in_use: {hidden: true}}.
...
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable. The full list is in cli.yaml under errors.
          example: device_in_use
        message: {type: string, description: 'What happened and why, in full'}
        hint: {type: [string, 'null'], description: 'What to do next, often an exact command'}
        docs_url: {type: [string, 'null'], format: uri}
        request_id: {type: string}
        details: {type: object, additionalProperties: true}
...